refactor: linkbundle, link 패키지를 하나로 통합
