+++ src/main/java/com/seong/shoutlink/domain/link/LinkBundle.java
@@ -1,4 +1,4 @@
-package com.seong.shoutlink.domain.linkbundle;
+package com.seong.shoutlink.domain.link;
 
 import lombok.Getter;
