+++ src/main/java/com/seong/shoutlink/domain/link/repository/LinkBundleRepositoryImpl.java
@@ -1,10 +1,10 @@
-package com.seong.shoutlink.domain.linkbundle.repository;
+package com.seong.shoutlink.domain.link.repository;
 
 import com.seong.shoutlink.domain.hub.Hub;
-import com.seong.shoutlink.domain.linkbundle.HubLinkBundle;
-import com.seong.shoutlink.domain.linkbundle.LinkBundle;
-import com.seong.shoutlink.domain.linkbundle.MemberLinkBundle;
-import com.seong.shoutlink.domain.linkbundle.service.LinkBundleRepository;
+import com.seong.shoutlink.domain.link.HubLinkBundle;
+import com.seong.shoutlink.domain.link.LinkBundle;
+import com.seong.shoutlink.domain.link.MemberLinkBundle;
+import com.seong.shoutlink.domain.link.service.LinkBundleRepository;
 import com.seong.shoutlink.domain.member.Member;
 import java.util.List;
 import java.util.Optional;
