+++ src/main/java/com/seong/shoutlink/domain/link/service/LinkBundleRepository.java
@@ -1,9 +1,9 @@
-package com.seong.shoutlink.domain.linkbundle.service;
+package com.seong.shoutlink.domain.link.service;
 
 import com.seong.shoutlink.domain.hub.Hub;
-import com.seong.shoutlink.domain.linkbundle.HubLinkBundle;
-import com.seong.shoutlink.domain.linkbundle.LinkBundle;
-import com.seong.shoutlink.domain.linkbundle.MemberLinkBundle;
+import com.seong.shoutlink.domain.link.HubLinkBundle;
+import com.seong.shoutlink.domain.link.LinkBundle;
+import com.seong.shoutlink.domain.link.MemberLinkBundle;
 import com.seong.shoutlink.domain.member.Member;
 import java.util.List;
 import java.util.Optional;
