+++ src/main/java/com/seong/shoutlink/domain/link/service/LinkBundleService.java
@@ -1,19 +1,19 @@
-package com.seong.shoutlink.domain.linkbundle.service;
+package com.seong.shoutlink.domain.link.service;
 
 import com.seong.shoutlink.domain.exception.ErrorCode;
 import com.seong.shoutlink.domain.exception.ShoutLinkException;
 import com.seong.shoutlink.domain.hub.Hub;
 import com.seong.shoutlink.domain.hub.service.HubRepository;
 import com.seong.shoutlink.domain.hub.service.HubMemberRepository;
-import com.seong.shoutlink.domain.linkbundle.HubLinkBundle;
-import com.seong.shoutlink.domain.linkbundle.LinkBundle;
-import com.seong.shoutlink.domain.linkbundle.MemberLinkBundle;
-import com.seong.shoutlink.domain.linkbundle.service.request.CreateHubLinkBundleCommand;
-import com.seong.shoutlink.domain.linkbundle.service.request.FindHubLinkBundlesCommand;
-import com.seong.shoutlink.domain.linkbundle.service.request.FindLinkBundlesCommand;
-import com.seong.shoutlink.domain.linkbundle.service.response.CreateLinkBundleCommand;
-import com.seong.shoutlink.domain.linkbundle.service.response.CreateLinkBundleResponse;
-import com.seong.shoutlink.domain.linkbundle.service.response.FindLinkBundlesResponse;
+import com.seong.shoutlink.domain.link.HubLinkBundle;
+import com.seong.shoutlink.domain.link.LinkBundle;
+import com.seong.shoutlink.domain.link.MemberLinkBundle;
+import com.seong.shoutlink.domain.link.service.request.CreateHubLinkBundleCommand;
+import com.seong.shoutlink.domain.link.service.request.FindHubLinkBundlesCommand;
+import com.seong.shoutlink.domain.link.service.request.FindLinkBundlesCommand;
+import com.seong.shoutlink.domain.link.service.response.CreateLinkBundleCommand;
+import com.seong.shoutlink.domain.link.service.response.CreateLinkBundleResponse;
+import com.seong.shoutlink.domain.link.service.response.FindLinkBundlesResponse;
 import com.seong.shoutlink.domain.member.Member;
 import com.seong.shoutlink.domain.member.service.MemberRepository;
 import jakarta.annotation.Nullable;
