+++ src/main/java/com/seong/shoutlink/domain/link/service/request/FindHubLinkBundlesCommand.java
@@ -1,4 +1,4 @@
-package com.seong.shoutlink.domain.linkbundle.service.request;
+package com.seong.shoutlink.domain.link.service.request;
 
 import jakarta.annotation.Nullable;
