+++ src/main/java/com/seong/shoutlink/domain/link/service/request/FindLinkBundlesCommand.java
@@ -1,0 +1,5 @@
+package com.seong.shoutlink.domain.link.service.request;
+
+public record FindLinkBundlesCommand(Long memberId) {
+
+}
