+++ src/main/java/com/seong/shoutlink/domain/link/service/response/FindLinkBundlesResponse.java
@@ -1,6 +1,6 @@
-package com.seong.shoutlink.domain.linkbundle.service.response;
+package com.seong.shoutlink.domain.link.service.response;
 
-import com.seong.shoutlink.domain.linkbundle.LinkBundle;
+import com.seong.shoutlink.domain.link.LinkBundle;
 import java.util.List;
 
 public record FindLinkBundlesResponse(List<FindLinkBundleResponse> linkBundles) {
