--- src/main/java/com/seong/shoutlink/domain/linkbundle/service/request/FindLinkBundlesCommand.java
@@ -1,5 +1,0 @@
-package com.seong.shoutlink.domain.linkbundle.service.request;
-
-public record FindLinkBundlesCommand(Long memberId) {
-
-}
