+++ src/test/java/com/seong/shoutlink/domain/linkbundle/service/LinkBundleServiceTest.java
@@ -7,15 +7,16 @@ import com.seong.shoutlink.domain.exception.ErrorCode;
 import com.seong.shoutlink.domain.exception.ShoutLinkException;
 import com.seong.shoutlink.domain.hub.Hub;
 import com.seong.shoutlink.domain.hubmember.repository.StubHubMemberRepository;
-import com.seong.shoutlink.domain.linkbundle.LinkBundle;
+import com.seong.shoutlink.domain.link.service.LinkBundleService;
+import com.seong.shoutlink.domain.link.LinkBundle;
 import com.seong.shoutlink.domain.linkbundle.repository.StubLinkBundleRepository;
-import com.seong.shoutlink.domain.linkbundle.service.request.CreateHubLinkBundleCommand;
-import com.seong.shoutlink.domain.linkbundle.service.request.FindHubLinkBundlesCommand;
-import com.seong.shoutlink.domain.linkbundle.service.request.FindLinkBundlesCommand;
-import com.seong.shoutlink.domain.linkbundle.service.response.CreateLinkBundleCommand;
-import com.seong.shoutlink.domain.linkbundle.service.response.CreateLinkBundleResponse;
-import com.seong.shoutlink.domain.linkbundle.service.response.FindLinkBundleResponse;
-import com.seong.shoutlink.domain.linkbundle.service.response.FindLinkBundlesResponse;
+import com.seong.shoutlink.domain.link.service.request.CreateHubLinkBundleCommand;
+import com.seong.shoutlink.domain.link.service.request.FindHubLinkBundlesCommand;
+import com.seong.shoutlink.domain.link.service.request.FindLinkBundlesCommand;
+import com.seong.shoutlink.domain.link.service.response.CreateLinkBundleCommand;
+import com.seong.shoutlink.domain.link.service.response.CreateLinkBundleResponse;
+import com.seong.shoutlink.domain.link.service.response.FindLinkBundleResponse;
+import com.seong.shoutlink.domain.link.service.response.FindLinkBundlesResponse;
 import com.seong.shoutlink.domain.member.Member;
 import com.seong.shoutlink.domain.member.MemberRole;
 import com.seong.shoutlink.domain.member.repository.StubMemberRepository;
