+++ src/test/java/com/seong/shoutlink/global/client/GeminiClientTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.seong.shoutlink.domain.link.Link;
-import com.seong.shoutlink.domain.linkbundle.LinkBundle;
+import com.seong.shoutlink.domain.link.LinkBundle;
 import com.seong.shoutlink.domain.tag.service.ai.GenerateAutoTagCommand;
 import com.seong.shoutlink.domain.link.LinkBundleAndLinks;
 import com.seong.shoutlink.domain.tag.service.ai.GeneratedTag;
