+++ src/test/java/com/seong/shoutlink/global/event/DomainEventListenerTest.java
@@ -10,10 +10,10 @@ import com.seong.shoutlink.domain.link.repository.LinkEntity;
 import com.seong.shoutlink.domain.link.service.LinkService;
 import com.seong.shoutlink.domain.link.service.request.CreateLinkCommand;
 import com.seong.shoutlink.domain.link.service.response.CreateLinkResponse;
-import com.seong.shoutlink.domain.linkbundle.LinkBundle;
-import com.seong.shoutlink.domain.linkbundle.MemberLinkBundle;
-import com.seong.shoutlink.domain.linkbundle.repository.LinkBundleEntity;
-import com.seong.shoutlink.domain.linkbundle.service.LinkBundleRepository;
+import com.seong.shoutlink.domain.link.LinkBundle;
+import com.seong.shoutlink.domain.link.MemberLinkBundle;
+import com.seong.shoutlink.domain.link.repository.LinkBundleEntity;
+import com.seong.shoutlink.domain.link.service.LinkBundleRepository;
 import com.seong.shoutlink.domain.member.Member;
 import com.seong.shoutlink.domain.member.service.MemberRepository;
 import com.seong.shoutlink.fixture.DomainFixture.DomainFixture;
